rename clipboard2 to clipboard
