+++ crates/gpui/src/platform/linux/x11.rs
@@ -1,5 +1,5 @@
 mod client;
-mod clipboard2;
+mod clipboard;
 mod display;
 mod event;
 mod window;
